player_choice(): create function
This is a small-ish change to the program, but quite a big change to the
way of thinking about it.

Thinking ahead, we know we're going to be asking the player to choose
from a list of options a few times.  So we bundle this behaviour into a
'function' to save repeating ourselves.

http://www.codecademy.com/courses/python-beginner-c7VZg/0/1

We then use this new function to get the same behaviour we had before.