work out which list of numbers the player sees
Lists come into this twice:

One is the "n_letters_in_colour in [3, 5]" test, which asks whether the
value of n_letters_in_colour is anywhere in the list [3, 5].  This is a
shorter way of saying

    if n_letters_in_colour == 3 or n_letters_in_colour == 5:
        # More code here.

The second appearance of lists is that we assign the label
'number_choices' either to the list ['1', '2', '5', '6'] or to the list
['3', '4', '7', '8'], depending on whether the player chose an odd- or
even-lettered colour.

Finally note that we make the number choices be strings not actual
numbers.  This isn't utterly necessary as we have the code here, but
it's good to be consistent --- the colour names were strings, so it's
good to make the number choices strings too.