ensure player chooses from list of choices
In the player_choice() function, we want to keep going round and round
until the player chooses something in the list.  In other words, while
they have NOT chosen from the list, we keep asking.

http://www.codecademy.com/courses/python-beginner-en-cxMGf/0/1

Before the player has typed an answer, they certainly haven't given us
an OK answer, so we start with

    player_has_chosen_ok = False

Then after they've typed something in, we re-check whether it's OK:

    player_has_chosen_ok = (choice in choices)

The (choice in choices) gives True or False --- same idea as in

    if first_number_chosen in ['1', '5']:
        # Etc.

except we label the result (True or False) for later use instead of
immediately acting on it in the 'if' statement.

Now the player can't type nonsense into the game.